Fast-update
file structure and click call back for popupWithImage in Card.
