+++ webpack.config.js
@@ -5,7 +5,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
-    entry: { main: './src/scripts/script.js' },
+    entry: { main: './src/index.js' },
     output: {
         path: path.resolve(__dirname, 'dist'),
         filename: 'main.js',
